// src/components/_uikit/widgets/Menu/components/PanelBody.tsx
import React from "react";
import styled from "styled-components";
import { useLocation } from "react-router-dom";
import { Link } from "components/_uikit";
import { SvgProps } from "../../../components/Svg";
import * as IconModule from "../icons";
import Accordion from "./Accordion";
import { MenuEntry, LinkLabel, LinkStatus } from "./MenuEntry";
import MenuLink from "./MenuLink";
import { PanelProps, PushedProps } from "../types";

interface Props extends PanelProps, PushedProps {
  isMobile: boolean;
}

const Icons = IconModule as unknown as { [key: string]: React.FC<SvgProps> };

const Container = styled.div`
  display: flex;
  flex-direction: column;
  overflow-y: auto;
  overflow-x: hidden;
  height: 100%;
`;
const LogoContainer = styled.div`
  height: 89px;
  position: relative;
  background-color: ${({ theme }) => theme.colors.tertiary};
  img {
    
  }
  border-bottom: solid 1px ${({ theme }) => theme.colors.background};
`;
const Logo = styled.img`
  height: 31px;
  position: absolute;
  top: 29px;
  left: 24px;   
`;
const LogoText = styled.img`
  height: 31px;
  position: absolute;
  top: 29px;
  left: 44px;   
`;

const PanelBody: React.FC<Props> = ({ isPushed, pushNav, isMobile, links }) => {
  const location = useLocation();

  // Close the menu when a user clicks a link on mobile
  const handleClick = isMobile ? () => pushNav(false) : undefined;

  return (
    <Container onMouseEnter={() => {pushNav(true);}} onMouseLeave ={() => {pushNav(false);}}>
      <LogoContainer>
        <Link href="https://1tronic.io/" external>
          {isPushed ? (
            <div>
              <Logo src="images/1tronic/1tronic-symbol.svg" alt="Logo"/> 
              <LogoText src="images/1tronic/network-white.svg" alt="Logo"/>
            </div>
          ) : (
            <Logo src="images/1tronic/1tronic-symbol.svg" alt="Logo"/>
          )}
        </Link>
      </LogoContainer>
      {links.map((entry) => {
        const Icon = Icons[entry.icon];
        const iconElement = <Icon width="24px" mr="8px" />;
        const calloutClass = entry.calloutClass ? entry.calloutClass : undefined;

        if (entry.items) {
          const itemsMatchIndex = entry.items.findIndex((item) => item.href === location.pathname);
          const initialOpenState = entry.initialOpenState === true ? entry.initialOpenState : itemsMatchIndex >= 0;

          return (
            <Accordion
              key={entry.label}
              isPushed={isPushed}
              pushNav={pushNav}
              icon={iconElement}
              label={entry.label}
              status={entry.status}
              initialOpenState={initialOpenState}
              className={calloutClass}
              isActive={entry.items.some((item) => item.href === location.pathname)}
            >
              {isPushed &&
                entry.items.map((item) => (
                  <MenuEntry key={item.href} secondary isActive={item.href === location.pathname} onClick={handleClick}>
                    <MenuLink href={item.href}>
                      <LinkLabel isPushed={isPushed}>{item.label}</LinkLabel>
                      {item.status && (
                        <LinkStatus color={item.status.color} fontSize="14px">
                          {item.status.text}
                        </LinkStatus>
                      )}
                    </MenuLink>
                  </MenuEntry>
                ))}
            </Accordion>
          );
        }
        return (
          <MenuEntry key={entry.label} isActive={entry.href === location.pathname} className={calloutClass}>
            <MenuLink href={entry.href} onClick={handleClick}>
              {iconElement}
              <LinkLabel isPushed={isPushed}>{entry.label}</LinkLabel>
              {entry.status && (
                <LinkStatus color={entry.status.color} fontSize="14px">
                  {entry.status.text}
                </LinkStatus>
              )}
            </MenuLink>
          </MenuEntry>
        );
      })}
    </Container>
  );
};

export default PanelBody;
